+++ debug/eslint-plugin/test.jsx
@@ -1,0 +1,7 @@
+import PropTypes from "prop-types";
+
+const A = {};
+A.propTypes = { b: PropTypes.string };
+A.viewModelMeta = { b: "ignore", c: "int" };
+
+export default A;