+++ eslint-local-rules.js
@@ -1,0 +1,2 @@
+// NOTE: This enables local debugging of the eslint plugin
+module.exports = require("./eslint-plugin").rules;