+++ eslint-plugin/index.js
@@ -1,5 +1,6 @@
 module.exports = {
   rules: {
-    "no-errors": require("./rules/no-errors")
+    "no-errors": require("./rules/no-errors"),
+    "no-unused-meta": require("./rules/no-unused-meta")
   }
 };