+++ eslint-plugin/rules/no-unused-meta/index.js
@@ -1,0 +1,35 @@
+const t = require("@babel/types");
+
+const viewModelsVisitor = require("../../utils/view-models-visitor");
+
+const message = name => `'${name}' is not defined in component propTypes`;
+
+// TODO: add option for overriding the ".jsx" check
+module.exports = {
+  create: function(context) {
+    if (!context.getFilename().includes(".jsx")) {
+      return {};
+    }
+
+    // The exported visitor functions gather information about propTypes, viewModelMeta and file exports. The 'Program:exit' visitor will execute last. When it runs, the validate function is called with the gathered data. 'validate' will report any errors to eslint via the 'context' object.
+    return viewModelsVisitor(getState => ({
+      "Program:exit": () => {
+        const { metaTypes, propTypes } = getState();
+
+        if (!t.isObjectExpression(metaTypes)) return;
+        if (!t.isObjectExpression(propTypes)) {
+          context.report(metaTypes, "Component has no propTypes");
+          return;
+        }
+
+        const propNames = propTypes.properties.map(p => p.key.name);
+
+        metaTypes.properties.forEach(node => {
+          if (!propNames.includes(node.key.name)) {
+            context.report(node.key, message(node.key.name));
+          }
+        });
+      }
+    }));
+  }
+};