+++ CHANGELOG.md
@@ -3,8 +3,9 @@
 ## 1.0.0
 
 - BREAKING: [eslint plugin] Renames `all` rule to `no-errors`
-- Fixes eslint plugin not reporting errors on `PropTypes.any` and `PropTypes.symbol`
-- Fixes eslint plugin not reporting errors on identifiers in meta
+- FEATURE: [eslint plugin] Adds `no-unused-meta` rule
+- FIX: [eslint plugin] Fixes eslint plugin not reporting errors on `PropTypes.any` and `PropTypes.symbol`
+- FIX: [eslint plugin] Fixes eslint plugin not reporting errors on identifiers in meta
 
 ## 0.2.0